Fix: 성별은 여러번 수정 안되도록 변경, 기타 코드 리팩토링, 프로필 사진 오류 수정?
