+++ src/app/mypage/page.tsx
@@ -2,9 +2,9 @@ import Profile from '@/components/mypage/Profile';
 
 const MyPage = () => {
   return (
-    <div>
+    <>
       <Profile />
-    </div>
+    </>
   );
 };
