+++ src/components/mypage/Favorite.tsx
@@ -2,7 +2,7 @@ import { Chip, Select, SelectItem } from '@nextui-org/react';
 import { useEffect, useState } from 'react';
 import { favoriteOptions } from '@/utils/FavoriteData';
 import { clientSupabase } from '@/utils/supabase/client';
-import { IsEditingType } from '@/types/userTypes';
+import type { IsEditingType } from '@/types/userTypes';
 import { useGetUserDataQuery } from '@/hooks/useQueries/useUserQuery';
 import { useFavoriteStore } from '@/store/userStore';
 import { customErrToast } from '../common/customToast';
