+++ src/hooks/useMutation/useSchoolMutation.ts
@@ -1,5 +1,5 @@
 import { updateSchool } from '@/service';
-import { UpdateSchoolType } from '@/types/userTypes';
+import type { UpdateSchoolType } from '@/types/userTypes';
 import { useMutation } from '@tanstack/react-query';
 
 export const useSchoolUpdateMutation = () =>
