+++ src/hooks/useQueries/useUserQuery.ts
@@ -13,6 +13,11 @@ export const useGetUserDataQuery = () => {
 
   const isLoggedIn = data?.user_id ? true : false;
 
+  // 1) initialData
+  // 2) suspense + fallback UI 사용
+  // 3) 조건부 분기를 통한 fallback UI 사용
+  // 4) 프리패치 쿼리 (서버에서 유저에 대한 데이터를 들고 내려온다)
+
   return { data, isPending, isError, error, isLoggedIn };
 };
