+++ src/query/user/userQueryKeys.ts
@@ -1,3 +1,3 @@
-export const USER_DATA_QUERY_KEY = ['user_data'];
-export const USER_POST_QUERY_KEY = ['user_post'];
-export const USER_LIKE_POST_QUERY_KEY = ['user_like_post'];
+export const USER_DATA_QUERY_KEY = 'user_data';
+export const USER_POST_QUERY_KEY = 'user_post';
+export const USER_LIKE_POST_QUERY_KEY = 'user_like_post';
